# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Wallet API Documentation",
        "description": "Hệ thống REST API Ví Điện Tử (E-Wallet) - Laravel 11",
        "contact": {
            "email": "admin@e-wallet.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "E-Wallet API Server"
        }
    ],
    "paths": {
...
        "/api/v1/wallet/balance": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Kiểm tra số dư ví hiện tại",
                "description": "Trả về thông tin ví của người dùng đang được đăng nhập gồm số dư và tiền tệ.",
                "operationId": "getWalletBalance",
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "wallet_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "balance": {
                                                    "type": "string",
                                                    "example": "1500000.00"
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "VND"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/wallet/deposit": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Webhook nhận nạp tiền",
                "description": "API Webhook dành cho các cổng thanh toán bên thứ ba gửi IPN để cộng tiền vào tài khoản ví sau khi thanh toán thành công.",
                "operationId": "depositWallet",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "wallet_id",
                                    "amount",
                                    "reference_id",
                                    "signature"
                                ],
                                "properties": {
                                    "wallet_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500000
                                    },
                                    "reference_id": {
                                        "type": "string",
                                        "example": "dep_ref_982347923"
                                    },
                                    "signature": {
                                        "type": "string",
                                        "example": "9e107d9d372bb6826bd81..."
                                    },
                                    "metadata": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nạp tiền thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Nạp tiền vào tài khoản thành công."
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction_id": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "balance_after": {
                                                    "type": "string",
                                                    "example": "500000.00"
                                                },
                                                "reference_id": {
                                                    "type": "string",
                                                    "example": "dep_ref_982347923"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/wallet/transfer": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Chuyển tiền giữa hai ví",
                "description": "Trừ số dư ví A (người chuyển hiện tại), cộng số dư ví B, ghi nhận 2 giao dịch transfer_out và transfer_in.",
                "operationId": "transferMoney",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "receiver_wallet_id",
                                    "amount"
                                ],
                                "properties": {
                                    "receiver_wallet_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50000
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Trả tiền ăn trưa"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chuyển tiền thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Chuyển tiền thành công."
                                        },
                                        "data": {
                                            "properties": {
                                                "reference_id": {
                                                    "type": "string",
                                                    "example": "uuid-string-xxxx-yyyy"
                                                },
                                                "balance": {
                                                    "type": "string",
                                                    "example": "1450000.00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}